# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "枢衡 Fulcrum API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/eval/report": {
      "get": {
        "summary": "Eval Report",
        "operationId": "eval_report_eval_report_get",
        "parameters": [
          {
            "name": "fulcrum_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fulcrum Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EvalReportDTO"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Eval Report Eval Report Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eval/run": {
      "post": {
        "summary": "Run Eval",
        "operationId": "run_eval_eval_run_post",
        "parameters": [
          {
            "name": "fulcrum_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fulcrum Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EvalReportDTO"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Run Eval Eval Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Chat Completions",
        "operationId": "chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/call": {
      "post": {
        "summary": "Tools Call",
        "operationId": "tools_call_tools_call_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gateway/chat": {
      "post": {
        "summary": "Gateway Chat",
        "description": "前置网关:判恶意 → 拦截/审核/放行;仅放行时转发企业智能体并回传其真实回复。",
        "operationId": "gateway_chat_gateway_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "fulcrum-demo"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tool_calls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeDTO"
            },
            "type": "array",
            "title": "Outcomes"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "content"
        ],
        "title": "ChatResponse"
      },
...
      "Disposition": {
        "type": "string",
        "enum": [
          "allow",
          "sanitize",
          "approve",
          "block"
        ],
        "title": "Disposition",
        "description": "分级处置(也复用于供应链评级:allow/review<=sanitize/approve/block)。"
      },
      "EvalMetricsDTO": {
        "properties": {
          "asr_baseline": {
            "type": "number",
            "title": "Asr Baseline",
            "default": 0.0
          },
          "asr_fulcrum": {
            "type": "number",
            "title": "Asr Fulcrum",
            "default": 0.0
          },
          "asr_reduction": {
            "type": "number",
            "title": "Asr Reduction",
            "default": 0.0
          },
          "recall_bsr": {
            "type": "number",
            "title": "Recall Bsr",
            "default": 0.0
          },
          "precision": {
            "type": "number",
            "title": "Precision",
            "default": 0.0
          },
          "fpr": {
            "type": "number",
            "title": "Fpr",
            "default": 0.0
          },
          "utility": {
            "type": "number",
            "title": "Utility",
            "default": 0.0
          },
          "decision_accuracy": {
            "type": "number",
            "title": "Decision Accuracy",
            "default": 0.0
          },
          "audit_complete_rate": {
            "type": "number",
            "title": "Audit Complete Rate",
            "default": 0.0
          },
          "hash_chain_pass_rate": {
            "type": "number",
            "title": "Hash Chain Pass Rate",
            "default": 0.0
          },
          "totals": {
            "$ref": "#/components/schemas/EvalTotalsDTO"
          }
        },
        "type": "object",
        "title": "EvalMetricsDTO"
      },
      "EvalReportDTO": {
        "properties": {
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "metrics": {
            "$ref": "#/components/schemas/EvalMetricsDTO"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/EvalSampleDTO"
            },
            "type": "array",
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "dataset",
          "metrics"
        ],
        "title": "EvalReportDTO"
      },
      "EvalSampleDTO": {
        "properties": {
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "attack_type": {
            "type": "string",
            "title": "Attack Type"
          },
          "malicious": {
            "type": "boolean",
            "title": "Malicious"
          },
          "expected": {
            "type": "string",
            "title": "Expected"
          },
          "predicted": {
            "type": "string",
            "title": "Predicted"
          },
          "held": {
            "type": "boolean",
            "title": "Held"
          },
          "attack_succeeded": {
            "type": "boolean",
            "title": "Attack Succeeded"
          },
          "decision_correct": {
            "type": "boolean",
            "title": "Decision Correct"
          },
          "audit_ok": {
            "type": "boolean",
            "title": "Audit Ok",
            "default": true
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "sample_id",
          "attack_type",
          "malicious",
          "expected",
          "predicted",
          "held",
          "attack_succeeded",
          "decision_correct"
        ],
        "title": "EvalSampleDTO"
      },
      "EvalTotalsDTO": {
        "properties": {
          "samples": {
            "type": "integer",
            "title": "Samples",
            "default": 0
          },
          "malicious": {
            "type": "integer",
            "title": "Malicious",
            "default": 0
          },
          "benign": {
            "type": "integer",
            "title": "Benign",
            "default": 0
          },
          "tp": {
            "type": "integer",
            "title": "Tp",
            "default": 0
          },
          "fn": {
            "type": "integer",
            "title": "Fn",
            "default": 0
          },
          "fp": {
            "type": "integer",
            "title": "Fp",
            "default": 0
          },
          "tn": {
            "type": "integer",
            "title": "Tn",
            "default": 0
          }
        },
        "type": "object",
        "title": "EvalTotalsDTO"
      },
...
      "GatewayChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": "anon"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GatewayChatRequest"
      },
      "GatewayChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "decision": {
            "$ref": "#/components/schemas/Disposition"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "forwarded": {
            "type": "boolean",
            "title": "Forwarded"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/GatewayFindingDTO"
            },
            "type": "array",
            "title": "Findings"
          },
          "reply": {
            "type": "string",
            "title": "Reply",
            "default": ""
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tools"
          },
          "upstream_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Error"
          },
          "output_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Disposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "output_risk_level": {
            "type": "string",
            "title": "Output Risk Level",
            "default": ""
          },
          "output_reason": {
            "type": "string",
            "title": "Output Reason",
            "default": ""
          },
          "output_blocked": {
            "type": "boolean",
            "title": "Output Blocked",
            "default": false
          },
          "output_sanitized": {
            "type": "boolean",
            "title": "Output Sanitized",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "decision",
          "risk_level",
          "forwarded",
          "reason",
          "max_score"
        ],
        "title": "GatewayChatResponse"
      },
...
      "GatewayFindingDTO": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "matched": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "score"
        ],
        "title": "GatewayFindingDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutcomeDTO": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "decision": {
            "$ref": "#/components/schemas/Disposition"
          },
          "executed": {
            "type": "boolean",
            "title": "Executed"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "tool_name",
          "decision",
          "executed"
        ],
        "title": "OutcomeDTO"
      },
...
      "ToolCallRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "tool_name"
        ],
        "title": "ToolCallRequest"
      },
      "ToolCallResponse": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Disposition"
          },
          "executed": {
            "type": "boolean",
            "title": "Executed"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "executed"
        ],
        "title": "ToolCallResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}